% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_to_math}
\alias{convert_to_math}
\title{Convert String to Math Expression}
\usage{
convert_to_math(x)
}
\arguments{
\item{x}{A string}
}
\description{
For the full range of possible math expressions
that will be converted see \link[grDevices]{plotmath}
}
