% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{+.default}
\alias{+.default}
\title{Default plus}
\usage{
\method{+}{default}(e1, e2)
}
\arguments{
\item{e1}{Add e1}

\item{e2}{To e2}
}
\description{
The default plus (primitve) if
it does not handle strings
}
