% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{str_to_math}
\alias{str_to_math}
\title{Convert String to Math Expression}
\usage{
str_to_math(char = "*")
}
\arguments{
\item{char}{}
}
\value{
Returns a function that turns a string into
an \link[base]{expression}. This can then
be used as a labeler function in ggplot2.
}
\description{
First, all whitespace is replace by the specified
string (star by default)
and then the string is converted to an expression.
For the full range of possible math expressions
that will be converted see \link[grDevices]{plotmath}
}
\examples{
df <- tibble(
  x = c("delta+1", "Delta AB", "alpha"),
  y = rep(1, length(x))
)

ggplot(df, aes(x, y)) +
  geom_point() +
  scale_x_discrete(labels = str_to_math())
}
